+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add support for #emit_yield in applied component (#4)
+
 ## 0.13 2022-01-19
 
 - Add support for partial parameter application (#3)